+++ phone/src/apps/messages/components/MessagesApp.tsx
@@ -72,7 +72,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 2,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
@@ -83,7 +83,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 3,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
@@ -94,7 +94,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 4,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
@@ -105,7 +105,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 5,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
@@ -116,7 +116,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 6,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
@@ -127,7 +127,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 7,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
@@ -138,7 +138,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 8,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
@@ -149,7 +149,7 @@ InjectDebugData([
         updatedAt: Date.now(),
       },
       {
-        id: 1,
+        id: 9,
         message: 'sup!',
         user_identifier: '1234',
         phone_number: '444-4444',
