+++ phone/src/os/apps/components/AppRoute.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Route } from 'react-router-dom';
+import { AppWithStartup } from './AppWithStartup';
+
+export const AppRoute = ({ id, component: Component, ...rest }) => {
+  return (
+    <Route
+      {...rest}
+      render={() => (
+        <AppWithStartup id={id}>
+          <Component />
+        </AppWithStartup>
+      )}
+    />
+  );
+};
